Rename conversation-handlers.tsx to efnyc-conversation-handlers.tsx.
